+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The Search page restores the previous results when returning to it
 - Temporal extents were incomplete in non-English languages
+- Don't show inline sort and filter options in Collection Search
+- Preserve sort in Collection Search
 
 ## [5.0.0] - 2026-07-31